inline & update endpoint item docs
